% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.direc.R
\name{equating.direc}
\alias{equating.direc}
\title{Equating IRT menggunakan equateIRT::direc}
\usage{
equating.direc(fd, threshold = 0.5)
}
\arguments{
\item{fd}{Folder yang berisi file parameter item.}

\item{threshold}{Batas selisih parameter untuk purification anchor.}
}
\value{
List: hasil[\link{1}] rangkuman, hasil[\link{2}] hasil detail.
}
\description{
Equating IRT menggunakan equateIRT::direc
}
