% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.plink.R
\name{equating.plink}
\alias{equating.plink}
\title{Three-Year IRT Equating using PLINK}
\usage{
equating.plink(fd, threshold = 0.5)
}
\arguments{
\item{fd}{Folder containing item parameter files.}

\item{threshold}{Maximum absolute difference in item difficulty
allowed during anchor purification.}
}
\value{
A list containing summary and detailed equating results.
}
\description{
Performs IRT equating between 2024-2025 and 2024-2026
using Mean/Mean, Mean/Sigma, Haebara, and Stocking-Lord methods.
}
