% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.sirt.R
\name{equating.sirt}
\alias{equating.sirt}
\title{Equating parameter SIRT untuk tiga tahun}
\usage{
equating.sirt(fd)
}
\arguments{
\item{fd}{Folder yang berisi file parameter.}
}
\value{
List dengan dua elemen:
\describe{
\item{hasil[\link{1}]}{Rangkuman hasil equating.}
\item{hasil[\link{2}]}{Hasil lengkap equating.}
}
}
\description{
Melakukan equating parameter 2024-2025 dan 2024-2026
berdasarkan kombinasi jenjang dan mata pelajaran.
}
