% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.direc.R
\name{equating.direc}
\alias{equating.direc}
\title{Equating IRT menggunakan equateIRT::direc}
\usage{
equating.direc(fd, threshold = 0.5, purification = TRUE)
}
\arguments{
\item{fd}{Folder yang berisi file parameter item.}

\item{threshold}{Batas praktis displacement (logit) untuk purification anchor.}

\item{purification}{TRUE = buang item anchor yang displace (2*SE & > threshold);
FALSE = pakai semua common item tanpa purifikasi.}
}
\value{
List: hasil[\link{1}] rangkuman, hasil[\link{2}] list.anchor (gabungan item per pasangan),
hasil[\link{3}] hasil detail.
}
\description{
Equating IRT menggunakan equateIRT::direc
}
