% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.sirt.R
\name{equating.sirt}
\alias{equating.sirt}
\title{Equating IRT menggunakan paket sirt}
\usage{
equating.sirt(fd, threshold = 0.5, purification = TRUE)
}
\arguments{
\item{fd}{Karakter. Path folder yang berisi file parameter item (.csv).}

\item{threshold}{Numerik. Batas absolut displacement (logit) untuk purifikasi anchor. Default 0.5.}

\item{purification}{Logikal. Jika TRUE, item dengan displacement > 2*SE dan > threshold akan dibuang.}
}
\value{
List berisi:
\describe{
\item{rangkuman}{Data frame ringkasan konstanta B (Mean-Mean, Haebara, Stocking-Lord) serta indikator kualitas penyetaraan (SD, Var, linkerror).}
\item{list.anchor}{List data frame item anchor sebelum purifikasi untuk setiap kombinasi.}
\item{detail}{List berisi detail hasil per pasangan tahun (24-25 dan 24-26), termasuk tabel displacement dan objek model \code{sirt}.}
}
}
\description{
Menghitung konstanta equating (B) untuk pasangan tahun 2024-2025 dan 2024-2026
menggunakan metode Rasch Equating dari paket \code{sirt}. Fungsi ini juga menyediakan
statistik deskriptif seperti Standar Deviasi, Varians, dan Link Error.
}
