% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.sns.R
\name{equating.sns}
\alias{equating.sns}
\title{Three-Year IRT Equating with Anchor Item Purification (SNSequate)}
\usage{
equating.sns(fd, threshold = 0.5, purification = TRUE)
}
\arguments{
\item{fd}{Folder path containing item parameter files.}

\item{threshold}{Maximum displacement threshold allowed for anchor items.}

\item{purification}{Logical. Perform anchor item purification if TRUE.}
}
\value{
A list containing summary table, anchor details, and raw irt.link models.
}
\description{
Three-Year IRT Equating with Anchor Item Purification (SNSequate)
}
